fix: address PR 171 follow-up review
